+++ src/main/java/com/girigiri/kwrental/reservation/controller/ReservationController.java
@@ -13,8 +13,8 @@ import org.springframework.web.bind.annotation.RestController;
 import com.girigiri.kwrental.auth.argumentresolver.Login;
 import com.girigiri.kwrental.auth.domain.SessionMember;
 import com.girigiri.kwrental.common.exception.BadRequestException;
+import com.girigiri.kwrental.reservation.dto.request.AddEquipmentReservationRequest;
 import com.girigiri.kwrental.reservation.dto.request.AddLabRoomReservationRequest;
-import com.girigiri.kwrental.reservation.dto.request.AddReservationRequest;
 import com.girigiri.kwrental.reservation.dto.response.RelatedReservationsInfoResponse;
 import com.girigiri.kwrental.reservation.dto.response.UnterminatedEquipmentReservationsResponse;
 import com.girigiri.kwrental.reservation.dto.response.UnterminatedLabRoomReservationsResponse;
@@ -32,15 +32,15 @@ public class ReservationController {
 
 	@PostMapping
 	public ResponseEntity<?> reserve(@Login final SessionMember sessionMember,
-		@RequestBody final AddReservationRequest addReservationRequest) {
-		reservationService.reserve(sessionMember.getId(), addReservationRequest);
+		@RequestBody final AddEquipmentReservationRequest addReservationRequest) {
+		reservationService.reserveEquipment(sessionMember.getId(), addReservationRequest);
 		return ResponseEntity.created(URI.create("/api/reservations")).build();
 	}
 
 	@PostMapping("/labRooms")
 	public ResponseEntity<?> reserveLabRoom(@Login final SessionMember sessionMember,
 		@RequestBody final AddLabRoomReservationRequest addLabRoomReservationRequest) {
-		final Long id = reservationService.reserve(sessionMember.getId(), addLabRoomReservationRequest);
+		final Long id = reservationService.reserveLabRoom(sessionMember.getId(), addLabRoomReservationRequest);
 		return ResponseEntity.created(URI.create("/api/reservations/" + id)).build();
 	}
