+++ src/main/java/com/girigiri/kwrental/reservation/dto/request/AddEquipmentReservationRequest.java
@@ -3,7 +3,7 @@ package com.girigiri.kwrental.reservation.dto.request;
 import lombok.Builder;
 
 @Builder
-public record AddReservationRequest(
+public record AddEquipmentReservationRequest(
 	String renterName,
 	String renterPhoneNumber,
 	String renterEmail,
